+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.4.1.0
-* Added `toString` convenience function for converting `Builder`s to `String`s without surrounding them with double quotes (as `show` does).
+* Added the utility functions `toStrictText` and `toString` for working with `Builder`s.
 
 # 0.4
 * Due to [GHC bug #5289](http://ghc.haskell.org/trac/ghc/ticket/5289), projects that depend on the `double-conversion` library (such as `text-format`, a dependency of `text-show`) may break due to GHC incorrectly linking against libstdc++. Therefore, `text-show` was changed so that it does not depend on `text-format` by default. This behavior can be changed by using the `-ftext-format` flag when using `cabal`.